+++ src/components/Logo.tsx
@@ -1,6 +1,8 @@
+import React from "react"
 import "./Logo.css"
 import logo from "../images/logo.svg"
-const Logo = () => {
+
+const Logo: React.FC = () => {
   return (
     <div className="logo_container">
         <img src={logo} alt="logo" className="bg_logo" />
